+++ apps/client/src/app/pages/portfolio/activities/create-or-update-activity-dialog/interfaces/interfaces.ts
@@ -1,9 +1,8 @@
 import { Activity, User } from '@ghostfolio/common/interfaces';
-
-import { Account } from '@prisma/client';
+import { AccountWithPlatform } from '@ghostfolio/common/types';
 
 export interface CreateOrUpdateActivityDialogParams {
-  accounts: Account[];
+  accounts: AccountWithPlatform[];
   activity: Partial<Omit<Activity, 'assetProfile' | 'id' | 'unitPrice'>> & {
     assetProfile: Activity['assetProfile'] | null;
     id: string | null;